try to add siglip text encoder

+++ llava/model/multimodal_encoder/siglip_encoder.py
@@ -794,7 +794,7 @@ class SigLipVisionTower(nn.Module):
             tokenizer = AutoTokenizer.from_pretrained("google/siglip-base-patch16-224")
             text_inputs = tokenizer(text, padding="max_length",
                                return_tensors="pt")
-            text_forward_out = self.vision_text(**text_inputs.to(device=self.device, dtype=self.dtype), output_hidden_states=True)
+            text_forward_out = self.vision_text(**text_inputs.to(device=self.device), output_hidden_states=True)
             last_hidden_state = text_forward_out.last_hidden_state
             pooled_output = text_forward_out.pooler_output
         return image_features, last_hidden_state
